requirements.txt
dropbox==10.1.1
html5lib==1.0.1
pyyaml==5.3.1
retrying==1.3.3
streamlink==2.0.0
tzlocal==2.1
